fix: [Interface Parameter type changed to int]

+++ internal/entity/interface.go
@@ -2,5 +2,5 @@ package entity
 
 type OrderRepositoryInterface interface {
 	Save(order *Order) error
-	GetTotal() (float64, error)
+	GetTotal() (int, error)
 }
